+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/simplifications/MethodOrLoopContractCompiler.java
@@ -1,72 +1,184 @@
 package com.aws.jverify.verifier.compiler.simplifications;
 
 import com.aws.jverify.ContractException;
+import com.aws.jverify.Nullable;
 import com.aws.jverify.common.Common;
 import com.aws.jverify.generated.*;
 import com.aws.jverify.verifier.compiler.JavaToDafnyCompiler;
 import com.aws.jverify.verifier.compiler.JavaViolationException;
 import com.aws.jverify.verifier.compiler.MethodOrLoopContract;
+import com.aws.jverify.verifier.compiler.Reporter;
 import com.sun.tools.javac.tree.JCTree;
 import com.sun.tools.javac.tree.TreeInfo;
+import com.sun.tools.javac.tree.TreeMaker;
 import com.sun.tools.javac.tree.TreeTranslator;
+import com.sun.tools.javac.util.Context;
+import com.sun.tools.javac.util.List;
 
 import java.util.ArrayList;
-import java.util.List;
+import java.util.Set;
 
 /*
 Extracts contracts from constructor, method or loop bodies
+Does not support lambdas.
  */
 public class MethodOrLoopContractCompiler extends TreeTranslator {
-    public final JavaToDafnyCompiler compiler;
+    private final TreeMaker maker;
+    private final Reporter reporter;
+    private final JVerifyMaker jverifyMaker;
 
-    public MethodOrLoopContractCompiler(JavaToDafnyCompiler compiler) {
-        this.compiler = compiler;
+    private MethodOrLoopContractCompiler(Context context) {
+        context.put(MethodOrLoopContractCompiler.class, this);
+        this.maker = TreeMaker.instance(context);
+        this.reporter = Reporter.instance(context);
+        this.jverifyMaker = JVerifyMaker.instance(context);
     }
 
-    
-    public List<JCTree.JCStatement> extractContract(JCTree.JCStatement statement,
-                                                    MethodOrLoopContract contract,
-                                                    boolean allowFooter) {
-        var statements = statement instanceof JCTree.JCBlock block
-                ? block.getStatements()
-                : List.of(statement);
-        return extractContract(statements, contract, allowFooter);
+    public static JCTree.@Nullable JCBlock getImplementation(JCTree.JCMethodDecl method) {
+        JCTree.JCStatement second = method.body.getStatements().get(1);
+        if (second instanceof JCTree.JCBlock block) {
+            return block;
+        }
+        return null;
     }
     
-    private List<JCTree.JCStatement> extractContract(List<JCTree.JCStatement> statements,
-                                                     MethodOrLoopContract contract,
-                                                     boolean allowFooter) {
-        var superOrThis = getSuperOrThis(statements);
-        
-        var remainingStatements = statements;
-        if (superOrThis != null) {
-            remainingStatements = remainingStatements.subList(1, statements.size());
+    public static boolean hasImplementation(JCTree.JCMethodDecl method) {
+        return method.body != null && method.body.getStatements().get(1) instanceof JCTree.JCBlock;
+    }
+
+    public static MethodOrLoopContractCompiler instance(Context context) {
+        MethodOrLoopContractCompiler instance = context.get(MethodOrLoopContractCompiler.class);
+        if (instance == null) {
+            instance = new MethodOrLoopContractCompiler(context);
         }
-        remainingStatements = addHeaderContracts(remainingStatements, contract);
-        if (allowFooter) {
-            remainingStatements = addFooterContracts(remainingStatements, contract);
+        return instance;
+    }
+    
+    public Set<JCTree.JCCompilationUnit> transform(Set<JCTree.JCCompilationUnit> envs) {
+        for (var env : envs) {
+            translate(env);
         }
-        var result = new ArrayList<>(remainingStatements);
-        if (superOrThis != null) {
-            result.addFirst(superOrThis);
+        return envs;
+    }
+    
+    public List<JCTree.JCStatement> extractContract(JavaToDafnyCompiler compiler, 
+                                                              JCTree.JCBlock block,
+                                                              MethodOrLoopContract contract) {
+        if (block.getStatements().size() != 2) {
+            throw new RuntimeException("Method body is not in contract + implementation format");
+        }
+        var contracts = (JCTree.JCBlock)block.getStatements().get(0);
+        var implementation = block.getStatements().get(1);
+        var hasImplementation = implementation instanceof JCTree.JCBlock;
+        for(var contractStatement : contracts.getStatements()) {
+            handleStatement(compiler, contractStatement, contract);
         }
 
+        List<JCTree.JCStatement> implementationStatements = hasImplementation 
+                ? ((JCTree.JCBlock) implementation).getStatements() 
+                : List.nil();
         if (contract.isPure) {
-            if (result.size() == 1) {
-                var statement = result.getFirst();
-                var bodyless = statement instanceof JCTree.JCThrow throwStatement &&
-                        throwStatement.expr.type.tsym.getQualifiedName().contentEquals(ContractException.class.getCanonicalName());
-                if (bodyless) {
-                    return result;
-                }
+            if (!implementationStatements.isEmpty()) {
+                contract.pureBody = compiler.expressionCompiler.toExpr(implementationStatements);
             }
-
-            if (!remainingStatements.isEmpty()) {
-                contract.pureBody = compiler.expressionCompiler.toExpr(remainingStatements);
-            }
-            return List.of();
+            return List.nil();
         }
-        return result;
+        return implementationStatements;
+    }
+    
+    @Override
+    public void visitTopLevel(JCTree.JCCompilationUnit tree) {
+        reporter.compilationUnit = tree;
+        super.visitTopLevel(tree);
+    }
+
+    @Override
+    public void visitDoLoop(JCTree.JCDoWhileLoop tree) {
+        if (tree.body != null) {
+            List<JCTree.JCStatement> newStatements = getNewStatements(tree, getStatements(tree.body), false);
+            tree.body = maker.Block(0, newStatements);
+        }
+        super.visitDoLoop(tree);
+    }
+
+    @Override
+    public void visitWhileLoop(JCTree.JCWhileLoop tree) {
+        if (tree.body != null) {
+            List<JCTree.JCStatement> newStatements = getNewStatements(tree, getStatements(tree.body), false);
+            tree.body = maker.Block(0, newStatements);
+        }
+        super.visitWhileLoop(tree);
+    }
+
+    @Override
+    public void visitForLoop(JCTree.JCForLoop tree) {
+        if (tree.body != null) {
+            List<JCTree.JCStatement> newStatements = getNewStatements(tree, getStatements(tree.body), false);
+            tree.body = maker.Block(0, newStatements);
+        }
+        super.visitForLoop(tree);
+    }
+    
+    List<JCTree.JCStatement> getStatements(JCTree.JCStatement statement) {
+        return statement instanceof JCTree.JCBlock block ? block.getStatements() : List.of(statement);
+    }
+
+    @Override
+    public void visitMethodDef(JCTree.JCMethodDecl tree) {
+        if (tree.body != null) {
+            var allowFooter = JavaToDafnyCompiler.isConstructor(tree.sym);
+            tree.body.stats = getNewStatements(tree, tree.body.getStatements(), allowFooter);
+        }
+        super.visitMethodDef(tree);
+    }
+
+    private List<JCTree.JCStatement> getNewStatements(JCTree tree, List<JCTree.JCStatement> statements, boolean allowFooter) {
+        var superOrThis = getSuperOrThis(statements);
+
+        if (superOrThis != null) {
+            statements = statements.tail;
+        }
+
+        java.util.List<JCTree.JCStatement> remainingStatements = new ArrayList<>(statements);
+        var contractStatements = new ArrayList<JCTree.JCStatement>();
+        remainingStatements = addHeaderContracts(remainingStatements, contractStatements);
+
+        if (allowFooter) {
+            int i;
+            for (i = remainingStatements.size() - 1; i > 0; i--) {
+                var current = remainingStatements.get(i);
+                boolean foundHeader = isContractStatement(current);
+                if (!foundHeader) {
+                    break;
+                }
+                contractStatements.add(current);
+            }
+            int footerContracts = i + 1;
+            remainingStatements = remainingStatements.subList(0, footerContracts);
+        }
+
+        maker.pos = tree.pos;
+        var contractBlock = maker.Block(0, List.from(contractStatements));
+
+        List<JCTree.JCStatement> implStatements;
+        if (superOrThis != null) {
+            remainingStatements = new ArrayList<>(remainingStatements);
+            remainingStatements.addFirst(superOrThis);
+        }
+        JCTree.JCBlock implementationBlock = maker.Block(0, List.from(remainingStatements));
+        implStatements = List.of(implementationBlock);
+        return implStatements.prepend(contractBlock);
+    }
+
+    private static boolean hasImplementation(java.util.List<JCTree.JCStatement> remainingStatements) {
+        boolean hasImplementation = true;
+        if (remainingStatements.size() == 1) {
+            var statement = remainingStatements.getFirst();
+            boolean isContractThrow = statement instanceof JCTree.JCThrow throwStatement &&
+                    throwStatement.expr.type.tsym.getQualifiedName().contentEquals(ContractException.class.getCanonicalName());
+            hasImplementation = !isContractThrow;
+        }
+        return hasImplementation;
     }
 
     private static JCTree.JCStatement getSuperOrThis(List<JCTree.JCStatement> statements) {
@@ -83,34 +195,58 @@ public class MethodOrLoopContractCompiler extends TreeTranslator {
         return superOrThis;
     }
 
-    private List<JCTree.JCStatement> addHeaderContracts(List<JCTree.JCStatement> statements,
-                                                        MethodOrLoopContract contract) {
-        int headerContracts;
+    private java.util.List<JCTree.JCStatement> addHeaderContracts(
+            java.util.List<JCTree.JCStatement> statements,
+            java.util.List<JCTree.JCStatement> headerContracts) {
         int i;
         for (i = 0; i < statements.size(); i++) {
             var statement = statements.get(i);
-            boolean foundHeader = handleStatement(compiler, statement, contract);
-            if (!foundHeader) {
+            boolean isContract = isContractStatement(statement);
+            if (!isContract) {
                 break;
             }
+            headerContracts.add(statement);
         }
-        headerContracts = i;
-        return statements.subList(headerContracts, statements.size());
+        return statements.subList(headerContracts.size(), statements.size());
     }
 
-    private List<JCTree.JCStatement> addFooterContracts(List<JCTree.JCStatement> statements, 
-                                                        MethodOrLoopContract contract) {
-        int i;
-        for (i = statements.size() - 1; i > 0; i--) {
-            var statement = statements.get(i);
-            boolean foundHeader = handleStatement(compiler, statement, contract);
-            if (!foundHeader) {
-                break;
+    private boolean isContractStatement(JCTree.JCStatement statement) {
+        if (!(statement instanceof JCTree.JCExpressionStatement expressionStatement
+                && expressionStatement.getExpression() instanceof JCTree.JCMethodInvocation invocation)) {
+            return false;
+        }
+        var jverifyMethod = JavaToDafnyCompiler.getJVerifyMethod(invocation);
+        if (jverifyMethod == null) {
+            return false;
+        }
+        
+        var methodName = jverifyMethod.getQualifiedName().toString();
+        switch (methodName) {
+            case "check" -> {
+                // not a header method, so stop here
+                return false;
+            }
+            case Common.PRECONDITION, "postcondition", "invariant", "decreases", "reads", "modifies" -> {
+                return true;
+            }
+            default -> {
+                reporter.reportError(invocation, "notSupported", methodName);
+                return false;
             }
         }
-        int footerContracts = i + 1;
-        return statements.subList(0, footerContracts);
     }
+
+    public JCTree.JCBlock getDefaultBody() {
+        JCTree.JCStatement implementation = jverifyMaker.contractThrow();
+        return maker.Block(0, List.of(
+                maker.Block(0, List.nil()),
+                implementation));
+    }
+
+    public static JCTree.JCBlock getContractBlock(JCTree.JCMethodDecl tree) {
+        return (JCTree.JCBlock) tree.body.getStatements().get(0);
+    }
+
 
     public static boolean handleStatement(JavaToDafnyCompiler compiler, JCTree.JCStatement statement, MethodOrLoopContract contract) {
         if (!(statement instanceof JCTree.JCExpressionStatement expressionStatement
@@ -197,12 +333,12 @@ public class MethodOrLoopContractCompiler extends TreeTranslator {
                     ? new ThisExpr(origin)
                     : new NameSegment(origin, NameCompiler.RETURN_VARIABLE_NAME, null);
             var origCondition = compiler.expressionCompiler.toExpr(lambda.getBody());
-            var condition = new LetExpr(origin, List.of(lhs), List.of(rhs), origCondition, true, null);
+            var condition = new LetExpr(origin, java.util.List.of(lhs), java.util.List.of(rhs), origCondition, true, null);
             header.postconditions.add(new AttributedExpression(condition, null, null));
 
         } else if (expr instanceof JCTree.JCMemberReference memberReference) {
             var origin = compiler.toOrigin(memberReference);
-            var argBindings = List.of(new ActualBinding(null,
+            var argBindings = java.util.List.of(new ActualBinding(null,
                     new NameSegment(origin, NameCompiler.RETURN_VARIABLE_NAME, null), false));
             var callee = new ExprDotName(origin,
                     compiler.expressionCompiler.toExpr(memberReference.expr),
